requirements.txt
# packaging
pip>=9
setuptools>=26
twine
wheel>=0.29

# data science
numpy
pandas

# tests/linter
black
coverage
flake8
pytest
yapf

# API
gcsfs
google-cloud-storage
mlflow
s3fs
requests

# utilities
six>=1.14
joblib
memoized-property
termcolor
datetime
python-dotenv

#WEB
streamlit
folium
plotly
treamlit_folium


